Rename _dom* files to dom*, since the non-standard naming is more annoying than helpful
