+++ lib/domDispose.ts
