+++ lib/domImpl.ts
@@ -1,5 +1,5 @@
-import {domDispose} from './_domDispose';
-import {attrsElem} from './_domMethods';
+import {domDispose} from './domDispose';
+import {attrsElem} from './domMethods';
 
 // Use the browser globals in a way that allows replacing them with mocks in tests.
 import {G} from './browserGlobals';
@@ -31,7 +31,7 @@ export interface IDomElementArgArray extends Array<DomElementArg> {}
 
 // The goal of the above declarations is to get help from TypeScript in detecting incorrect usage:
 // (See test/types/dom.ts for a test of this.)
-//  import {text, hide} from './_domMethods';
+//  import {text, hide} from './domMethods';
 //  dom('div', text('hello'));        // OK
 //  dom('div', hide(true));           // OK
 //  dom('div', {title: 'hello'});     // OK
